警告を表示するコードをpage本体からcomposablesに移行
